backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
requests==2.32.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.38
boto3==1.37.0